Split CWS's submission module as it was growing too large
