+++ cms/server/contest/submission/__init__.py
@@ -1,0 +1,50 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# Contest Management System - http://cms-dev.github.io/
+# Copyright © 2018 Luca Wehrstedt <luca.wehrstedt@gmail.com>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Package containing utilities for accepting submissions in CWS."""
+
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+from future.builtins.disabled import *  # noqa
+from future.builtins import *  # noqa
+
+from .check import get_submission_count, check_max_number, \
+    get_latest_submission, check_min_interval
+from .file_retrieval import ReceivedFile, InvalidArchive, \
+    extract_files_from_archive, extract_files_from_tornado
+from .file_matching import InvalidFilesOrLanguage, match_files_and_languages
+from .utils import fetch_file_digests_from_previous_submission, StorageFailed, \
+    store_local_copy
+
+
+__all__ = [
+    # check.py
+    "get_submission_count", "check_max_number", "get_latest_submission",
+    "check_min_interval",
+    # file_retrieval.py
+    "ReceivedFile", "InvalidArchive", "extract_files_from_archive",
+    "extract_files_from_tornado",
+    # file_matching.py
+    "InvalidFilesOrLanguage", "match_files_and_languages",
+    # utils.py
+    "fetch_file_digests_from_previous_submission", "StorageFailed",
+    "store_local_copy",
+]
